requirements.txt
markitdown==0.0.1a3
openai==1.58.1
pythondotenv==1.0.1